Modularized plugins, simplified shipkit java plugin
Reworked the plugins so that they are more composable and hence easier to use.
Finally, shipkit java plugin has no complexity, it only applies other plugins!!!
